Revert "Move theme provider into src dir"
This reverts commit 509d0ead6c966fdee776d2e6b0ae0ff3923b0dba.
