+++ theme-provider/index.ts
@@ -1,3 +1,1 @@
-export { useTheme, ThemeProvider } from './src/context/ThemeProvider';
-export { ThemeContextConfig } from './src/types/themeContextConfig';
-export { WhiteLabelColors } from './src/types/whiteLabelColors';
+export * from './context/ThemeProvider';
